Some changes to men editing system & add total count for roti.
