+++ fmb/users/emailroti.php
@@ -1,7 +1,4 @@
 <?php
-include('connection.php');
-include('getHijriDate.php');
-require_once '_sendMail.php';
 
 error_reporting(0);
 $day = date("l", strtotime("tomorrow"));
@@ -10,18 +7,28 @@ if ($day == 'Sunday') {
 	exit;
 }
 
-$tomorrow_date = date("Y-m-d", strtotime("+ 1 day"));
+if(isset($_GET['date'])) {
+	$tomorrow_date = $_GET['date'];		
+} else {
+	$tomorrow_date = date("Y-m-d", strtotime("+ 1 day"));
+}
 
 $extramsg = mysqli_query($link, "SELECT DISTINCT `Transporter` from thalilist WHERE Active = 1 AND extraRoti != 0 ORDER BY Transporter");
 while ($row_extra = mysqli_fetch_assoc($extramsg)) {
 	$sql = mysqli_query($link, "SELECT * from thalilist WHERE extraRoti != 0 AND `Transporter` LIKE '".$row_extra['Transporter']."'");
-	$msg .= "<b>" . $row_extra['Transporter'] . "</b><br/>";
+	$msgroti .= "<b>" . $row_extra['Transporter'] . "</b><br/>";
 	while ($row = mysqli_fetch_assoc($sql)) {
-		$msg .= "<b>".$row['extraRoti']." Extra Roti</b> - ";
-		$msg .= sprintf("%s - %s - %s - %s - %s - %s<br/>", $row['tiffinno'], $row['thalisize'], $row['NAME'], $row['CONTACT'], $row['wingflat'], $row['society']);
-		$msg .= '<br/>';
+		if($row['thalisize'] == 'Mini' || $row['thalisize'] == 'Small') {
+			$msgroti .= "<b>". 1 + $row['extraRoti'] ." Roti</b> - ";
+		} elseif($row['thalisize'] == 'Medium' || $row['thalisize'] == 'Large') {
+			$msgroti .= "<b>". 2 + $row['extraRoti'] ." Roti</b> - ";
+		} else {
+			$msgroti .= "<b>".$row['extraRoti']." Roti</b> - ";
+		}
+		$msgroti .= sprintf("%s - %s - %s - %s - %s - %s<br/>", $row['tiffinno'], $row['thalisize'], $row['NAME'], $row['CONTACT'], $row['wingflat'], $row['society']);
+		$msgroti .= '<br/>';
 	}
-	$msg .= '<br/>';
+	$msgroti .= '<br/>';
 }
 
 $transporter = mysqli_query($link, "SELECT DISTINCT `Transporter` from thalilist WHERE Active = 1 ORDER BY Transporter");
@@ -32,7 +39,7 @@ while ($row_trans = mysqli_fetch_assoc($transporter)) {
 
 $thaliSize = array();
 $hijridate = getHijriDate($tomorrow_date);
-$msg .= "<br/><b>Roti Count on $hijridate $day - $tomorrow_date</b><br/>";
+$msgroti .= "<br/><b>Roti Count on $hijridate $day - $tomorrow_date</b><br/>";
 $rotiTable = "<table border='1' ><tr><td style='padding: 2px 10px 2px 10px;'>Size</td>";
 foreach ($transporters as $transporter) {
 	$totalCount = 0;
@@ -66,11 +73,20 @@ foreach ($thaliSize as $size => $sizeCount) {
 		$rotiTable .= "<td style='padding: 2px 10px 2px 10px;'>" . $sizeCount[$transporter] . "</td>";
 	}
 	$rotiTable .= "<td style='padding: 2px 10px 2px 10px;'>".$totalSizeCount."</td></tr>";
-}
+}	
 
 $rotiTable .= "</table>";
 
-$msg .= $rotiTable;
+$msgroti .= $rotiTable;
+
+$totalCount = $totalSizeCount*4;
+
+$msgroti .= "<br/><b>Total Roti Count is $totalCount</b>";
 
 // send email
-sendEmail('kalimimohallapoona@gmail.com', 'Roti update ' . $tomorrow_date, $msg, null, null, true);
+sendEmail('kalimimohallapoona@gmail.com', 'Roti update ' . $tomorrow_date, $msgroti, null, null, true);
+
+if(isset($_GET['date'])) {
+	header("Location: /fmb/users/usermenu.php?action=send&date=" . $_GET['date']);	
+}
+
