+++ fmb/users/thalisearch.php
@@ -204,21 +204,19 @@ if (isset($_GET['year'])) {
 
   <?php $adminmenu = mysqli_query($link, "SELECT * FROM menu_list WHERE `menu_date` >= '" . date('Y-m-d') . "' AND `menu_type` = 'thaali' order by `menu_date` DESC") or die(mysqli_error($link));
   while ($amenu_values = mysqli_fetch_assoc($adminmenu)) {
+    $menu_id = $amenu_values['id'];
+    $menu_date = $amenu_values['menu_date'];
     if( !empty($values['Thali'])) {
       $adminumenu = mysqli_query($link, "SELECT * FROM user_menu WHERE `menu_date` = '".$amenu_values['menu_date']."' AND `thali` = '".$values['Thali']."'") or die(mysqli_error($link));
     }
     if(isset($adminumenu) && $adminumenu->num_rows > 0) {
       $rowaumenu = $adminumenu->fetch_assoc();
-      echo $menu_id = $rowaumenu['id'];
-      $menu_date = $rowaumenu['menu_date'];
       $menu_item = unserialize($rowaumenu['menu_item']);
       if( !empty($menu_item['roti']['qty'])) {
         $roti_qty = $menu_item['roti']['qty'];
       }
       $target = 'adminusermenu-'.$menu_id; 
     } else {
-      $menu_id = $amenu_values['id'];
-      $menu_date = $amenu_values['menu_date'];
       $menu_item = unserialize($amenu_values['menu_item']);
       if( !empty($values['thalisize'])) {
         if($values['thalisize'] == 'Mini') {
